Fixing treatment of hanging nodes
In 3D, we can have a situation where DoFs along an edge are constrained
without any of the faces being constrained (Imagine a 2x2x1 mesh where
elements (1,0), (0,1) and (1,1) are refined. Two of the children of the
(1,1) element will then share an edge with the coarse (0,0) element, and
thus have constrained DoFs along that edge). This commit fixes that
case.